Remove an incorrect and unnecessary materialization scope tree hack
As part of the materialization work, I added a hack where we would
hoist a binding up out of a branched-but-not-fenced subnode. This
apparently fixed some materialization cases where WITH+tuple is being
used like FOR, but it also pretty clearly breaks cardinality inference
for other cases using WITH+tuple like FOR!

This hack doesn't seem to be needed anymore, though. Hopefully because
the *real* underlying cause was fixed.