Fix another BRANCH related cardinality inference issue (#3018)
When figuring out visibility, we want to look one fence level out from
the current fence. When the current scope is a fence (which it usually
is), this is the same as parent_fence, but if the current scope is a
branch, we need to go out twice.

(Complicating the existing story is we current grab the parent *branch*;
this was something I did for materialization reasons and I think it might
have been wrong?)

A knock-on effect of this is that I needed to fix final schemamech
constraint compilation to pass in singletons, since
test_constraints_ddl_16, which has an exclusive constraint that uses ??,
only worked because we were mis-inferring cardinalities.