Fix/update for the build guide.
The build guide previously failed do to numerous errors.

1. Cloning sub-modules.
It was not stated in the build guide (or anywhere else) that you need to clone all the sub-modules of the repository. This lead to failed builds with nonsensical error messages.

2. Small issue with paths.
Probably a typo but the line `cd ../edgedb` should be only `cd edgedb`.