Fix/update to the build guide (#3041)
The build guide previously failed do to numerous errors.

1. Cloning sub-modules.
It was not stated in the build guide (or anywhere else) that you need to clone all the sub-modules of the repository. This lead to failed builds with nonsensical error messages.

2. Small issue with paths.
Probably a typo but the line `cd ../edgedb` should be only `cd edgedb`.

* Fixed build issue (cannot be over 79)

There was an issue with our previous pull request as one of the lines in the rst file was longer than 79 characters. This is now fixed


Co-authored-by: strace me <87979263+kozmer@users.noreply.github.com>