Switch `client_idle_timeout` to milliseconds / int32. Bump it to 60s.
Since we're about to add more timeouts, some of which would be directly
mapped to Postgres native config settings which are in milliseconds,
we're switching our own client_idle_timeout to milliseconds too for
consistency.

Sine int16 isn't big enough for expressing timeouts in milliseconds we
bump it to int32.  That allows us to specify timeouts up to 2147483
seconds (596 hours).