Fix many of migration proposal rejection
One major issue was that the logic for determining what the proposed
change that was being rejected did not match the logic for computing
the proposed change. This is fixed by pulling the info about the
rejected change directly from last_proposed, which also has the
advantage of eliminating another schema diff.

The downside of this is that reject no longer rejects anything unless
DESCRIBE CURRENT MIGRATION AS JSON has been run in the current
transaction. This works fine with the CLI, though. If this is a
problem, we can reuse the logic from DESCRIBE AS JSON to reproduce it
when needed.

can_alter and friends also needed to be changed to get types directly
from the relevant objects, since they were getting confused by sclass
being Pointer (which corresponds to the type of the pointers refdict.)

I had to adjust the confidence generation to prevent emitting a 1.0
confidence for some of the (very not 1.0 confidence!) suggestions
in my tests.

I then fiddled around a bit to try to make some of the messages in
those cases more granular.