Explicitly check for the byte length of runstate dir
Most systems have a fairly low limit for the total length of a Unix
socket path, e.g. on macOS its 104 bytes, and on Linux it's 108.  Rather
than failing with an obscure error and a traceback, validate the
specified or implied runstate path directory with an assumption for the
maximum Unix socket name we use.