Fix issues when migrating between regular and computed pointers (#3098)
Currently we create columns for single computed pointers, which
are kept empty. These can cause trouble if they are required and
the pointer is computed. Also, data in it can persist if a pointer
goes from real to computed and back to real.

Additionally, while we do not create a link table for multi computed
pointers, we do not delete it if a pointer *becomes* multi, which
can cause trouble.

Make sure that we only have columns/tables for non-computed data.

Also, properly delete triggers when a multi property is deleted
or when it becomes computed.

One behavior change as part of this is that if a computed property
is altered into a non-computed property with a default value,
existing objects get that default value, rather than staying empty.
This seems clearly correct to me, and matches what creating a new
property does, but there was a test that tested the old behavior.

Fixes #2411.