Make sure `False AND {}` is `{}`
The `std::AND` operator is not `OPTIONAL`, so we cannot safely fall back
to SQL `AND`.  Use the same technique as we do for `OR`.  Technically,
we could formulate this to only use the altered implementation if any of
the operands is optional, but that is left as a future optimization.