Make "and" and "or" in GraphQL conform to Postgres truthtable.
In EdgeQL all operators handle {} the same way. If the operands are
required, then the result is {} if any of the inputs are {}. This
applies to boolean operators as well, so that `not`, `and`, `or` will
all produce {} as the result if any of the operands are {}. This case
can then be explicitly handled by adding `??` to the operands or the the
result.

In Postgres `False AND NULL` is `False` and `True OR NULL` is `True`.
In the general case this inconsistency can cause hard to debug subtle
issues.

In GraphQL we don't have arbitrary expressions, but rather a very
restricted subset of them appearing as "filter". In that particular
subset of usage it makes sense to use "short-circuiting" definition of
"and" and "or" operators because that best conforms to the expectations
in this use-case.