Make json output format do base64 encoding as well, use that to drive casts
This requires some various plumbing changes to make this work in nested cases.

Some logic in the IR side cast generation needed to be updated to not
include an inaccurate intermediate type when casting a bytes array
literal to json.

I think that much of the logic on the IR side for compiling json casts
from arrays and tuples can be excised (though probably with the
addition of a call to eta_expand), but I'm leaving that to a follow-up.