Fix more issues with injecting typenames into std:: objects
We need to be able to insert views for injection inside of WITH/FOR
bindings, so we generalize expr_exposed to have an intermediate
exposure level corresponding to bindings.

To make problems like this less obscure, we make them *worse* by
adding an assert that prevents injecting into non-view shapes in
general.