Fix false positive cycle detections in functions
The function body tracer is checking for references to functions and
then uses the name prefix to look for all overloaded function
definitions to include in the dependency list.  Unfortunately, the
prefix checking code simply uses `startswith`, which could grab
completely unrelated functions if they happen to have a name with
a given prefix.

Fixes: #3115