Fix generated SET TYPE in migration when a create is split into an alter (#3191)
When an ALTER LINK that changes type gets split off from a CREATE TYPE
into an ALTER TYPE, we may need to include a USING statement. The body
of the USING statement doesn't actually matter---it's only needed because
the DDL side can't figure out that the cast expr isn't needed.

Fixes #3183.