Don't have {} act as a wildcard in assert_query_result
For assert_query_result, the empty dict `{}` acts as a "wild card",
that will match anything (because it checks each element in the dict,
and so if the dict is empty, it doesn't look at the returned data at all).

In addition to tests intentionally using it as a wildcard, there are
lots of tests that wanted it to mean "null or empty set".

For the former, I put in the actual value or put in the type,
which acts as a wildcard but in a slight more typed way.

The latter I fixed.

Fixes #3196.