server: Remove custom resolve logic for "localhost"
The current code is manually resolving "localhost" and then tries
to (a) test if IPv6 is functional on the platform or not, and
(b) based on that knowledge filter the list of what "localhost"
resolves into.

The problem with that is that it's a hack which only works for
localhost. If the server is run with `-I my.domain` which resolves
into an IPv6 address, the current workaround wouldn't do anything
about it.

Besides, the whole approach feels weird and seems unnecessary. Let's
remove the workaround and see how the system behaves without it.
It's been a while (and multiple Python / uvloop versions) since we
added this, maybe we don't need it anymore. And if a problem arises,
maybe we'll be able to find a better way to fix it.