Filter NULL arguments out when calling non-strict functions/operators (#3200)
We have two related problems with a related solution:
 - Our implementation of the AND operator, which carefully
   ensures that we return NULL if either input is NULL,
   is a huge performance problem that prevents the use of indexes
 - User defined functions with implementations that aren't strict in
   their arguments can end up producing values when they shouldn't.

To fix this, we add NOT NULL filters when calling functions/operators
with potentially NULL values. To avoid doing this *too* much (and to
make it work more often in singleton mode), we add a notion of
function/operators having a strict implementation, which allows
the check to be skipped.

Defined functions with all required singleton arguments are marked as
strict on creation, and most built-in operators and functions that do 
"USING SQL <something>" are marked as strict. A validation check
checks that things with "USING SQL FUNCTION" are really strict.

Fixes some #3185 related issues.
Fixes #3195.