server: Be tolerant toward failures to bind server sockets
This makes it so that failures to resolve or bind to some of the
specified `listen_addresses` (and `-I` on command line) are not fatal
and are instead logged as warnings.  This is needed because of the
unfortunate reality of bad DNS and network stack configurations in the
wild, where even `localhost` might not be a reliable listen interface
specification.  A good example of this is Github actions where,
`localhost` resolves to `::1`, but IPv6 does not actually work.

This effectively restores pre-176dcd5a04493 behavior and goes further to
tolerate binding failures as long as at least one socket binds.  This
matches PostgreSQL behavior.