Avoid generating EXISTS where clauses for type filters
Since we don't run card inference on type rewrites, the where clause
cardinality always gets compiled as if it is multi, which it isn't.

Run card inference on type rewrites, and tweak compilation of FILTER
to be both more natural reading and to crash on this sort of bug
instead.