Fix some problems involving toplevel statements in schema queries
Queries that are specified in the schema for various reasons could
be bare expressions, without a SELECT statement or similar.

Some functions that operated on schema objects and did not have a SELECT
in their definition would fail to include the rewrite CTEs in the function
body, since that code only triggered from fini_stmt.

An initial attempt to fix this by *always* including a SELECT revealed
bugs in that direction, also! (And was an unsatisfying solution.)

We fix these issues in two parts:
 1. Make less things driven by detecting the toplevel_stmt.
    Explicitly finalize the toplevel.
 2. Register the top level as the path_scope for any singletons
    in a schema expression, which prevents extra scope levels
    from causing trouble in SET TYPE USING queries and the like.

I'm hoping to do some more followup cleanups to tear out more
`is_toplevel` code.