Have ordering produce properly sub-nested hierarchies
Currently when an inherited parent object (like a pointer or
constraint) is deleted, all of its descendants have their implicit
delete op attached directly to the parent delete.

This causes problems when the object is a constraint and the property
is referenced by some other computed prop, since that computed prop
will be rechecked in the inconsistent state of constraint being
half-deleted.

Fix this by making a sub-sub-object delete properly get attached to
the sub-object delete.

Fixes #3280.