Avoid materializing in a few unnecessary cases
We materialize computed pointers if they refer to bindings that are
visible from the pointer definition, since those bindings might not be
visible at the use sites, and so we need the materialization to make
sure the correct bindings get referred to.

But this is only necessary if the bindings *aren't* visible from the
use sites, so add that as a check.

I ran into this while thinking about how GROUP BY is going to probably
be implemented via desugaring, but it's an improvement separate from
that.