Add a `bag` type that tells assert_query_result to ignore order
assert_query_result currently supports using sets to ignore order,
but that doesn't work for objects, which can't be hashed or sorted.

There is a system for specifying a sort key for internal data, but it
is way clunkier than just saying we don't care about the order.

I converted some places that were using sort= to use this.