Fix performance of queries containing references to multi-links from DML
When compiling link references that have a DML statement as a source,
we normally produce an "overlay" UNION containing the existing link
relation as well as the effects of the mutation.  Unfortunately,
Postgres seems to cope badly with JOINs containing UNIONs which are
composed of a mix of RelRange and RangeSubselect, producing horrible
HashJoin plans scanning the entire UNION.  Luckily, there's a
straightforward workaround: push the JOIN condition into UNION operands
and simply CROSS JOIN the UNION laterally.  While this is theoretically
beneficial for all kinds of UNIONs, we restrict this optimization to overlay
UNIONs only to limit the possibility of breakage as not all  UNIONs are
guaranteed to have correct path namespace and translation maps set up.