Improve performance of `IN`
A fix in #2329 changed the implementation of `IN` from an `ALL`/`ANY`
sublink to a `[NOT] EXISTS` sublink, which is slower in the trivial
cases like `foo in array_unpack(array)`.  The original fix was motivated
by tuple property breakage, and fortunately there is a simpler syntactic
fix for opaque tuples that allows keeping `ALL`/`ANY` sublinks in all
cases.

This does regress one test with a deeply nested ad-hoc tuple, which
crashes Postgres with an arcane "could not identify a hash function for
type record"