Always include the definition context namespace in computable contexts
We need to include the *original* source namespace in our ctx
namespace when compiling computables. The current mechanism of trying
to look up in view_sets or failing that using the source namespace
from the computable use, but this was failing to find it in some cases
with FOR.

Fix this by instead directly pulling in the namespace from qlctx. The
inclusion of qlctx's namespace nicely allows us to ditch so later
logic as well.

Additionally we need to merge the namespace into *both* sides in
get_view_map_remapping, to handle cases like referencing a `FOR`
variable where the current ns doesn't get merged in.

Fixes #3323.