Convert the docs to use lower-case keywords consistently. (#3348)
* Convert the docs to use lower-case keywords consistently.

The client library examples are left as is. The queries appear inside
strings and so upper-case keywords may be desirable in that instance.

The :eql-statement: no longer requires the title to be all in upper-case.
The statement itself is normalized to be lower-case for purposes of
referencing.

* Update the `project init` examples.

Issue #3366

* Update function cheatsheets.

Use the expression to define functions and not a string.

Issue #3366

* Stop using a link to `detached`.

Use either `:eql:op:detached` or `:eql:kw:detached`. The keyword usage
points to the "with block" page, whereas the operator refers tot he
stdlib.

Issue #3366

* Update the autogeneted meta grammar file.

Issue #3366