Optimize optional compilation of simple paths
Currently we put an optional wrapper around optional arguments in all
cases, include ones such as
`SELECT Issue { te :=  .time_estimate ?? -1 }` in which it would be sound
to simply rely on time_estimate being stored as NULL in the database
and omit the optional wrapper.

We can omit the optional wrapper when the argument is a path in which
the LHS is visible and the pointer is a single non-computed property.