Reduce the number of calls to delta_schema in the migration happy path (#3397)
Get rid of all calls to schemas_are_equal except to check if a migration
is complete once we already think it should be. Also, distinguish between
the "we have nothing proposed" case and the "we have consumed all proposed
advice" case, which saves one schema diff at the end.

This brings the time to migrate to constraints.esdl down on my machine
from 78s to 12s.