Provide a more detailed message in constraint error details field
Currently, constraint violation errors include a nicely-formatted
message that is intended to be displayed to application end user, and
so don't include any extra technical details, such as the type on which
the constraint is defined.  This makes life harder on developers, so
include a more elaborate message in the `details` field:

    edgedb> insert City2 { name := 'sesh' };
    edgedb error: ConstraintViolationError: name violates exclusivity constraint
      Detail: value of property 'name' of object type 'default::City2' violates exclusivity constraint

It would also be nice to include the constraint definition and the
offending value, but that requires a more elaborate fix.

Fixes: #3522