Determine func/oper argument type modifiers in advance of compiling args (#3571)
This is based on a previous attempt, #2132.

This allows us to compile arguments with the correct fences in the
first place, instead of trying to fix them up after the fact.

The key insight here is that if a function argument could be either
OPTIONAL or SINGLETON, it is safe to treat it as optional. We can
then mark it as non-optional, when we finalize things. In obscure cases
(when a path in a spuriously-optional argument gets factored with another
fenced or optional node), paths may have escaped already, which might
lead to a unneeded wrapper, but common cases do the right thing.

This change allowed me to rip out a bunch nasty mechanisms for dealing
with OPTIONAL that I introduced during the first few months of my
employment and that I have been trying to kill ever since
(optional_count, path_log).

This is about code-size-neutral but the added code is boring plumbing
(to find typemods before fully typechecking) and the removed code is
fiddly hacks.

This also proves that it should be possible to precompute a scope tree
without needing to do move typechecking before scope tree inference.