Fix inheritance from enum types (#3593)
The main thing is that the subtypes ought to be domain types, not enum
types themselves.

THe bulk of the code, however, all has to deal with fixing the
_undo_everything/_redo_everything scheme in the presence of inherited
enum types. The inherited enums need to be able to be deleted
temporarily if an element is being dropped from the parent type, and
that needs to be properly ordered with respect to properties that
depend on it. We deal with this by pulling the core part of the
property type switcharoos into the main sorted loop instead of having
it be separate.

Fixes #3578.