Generate fresh ids for free shapes when possible
Free shapes generated from cross products are not guarenteed to have
unique ids, however. So while
  `for x in User union { asdf := x.name ++ "!", x := .id}`
generates unique ids,
  `select (User, { asdf := User.name ++ "!", x := .id})`
does not.

I'm not sure if there is a good way to avoid this. We ban cross products
of volatile computations for this reason, but doing that to free objects
seems aggressive?