Add an `FOR GROUP` construct and desugaring code from GROUP
The implementation strategy for our nice shape-based `GROUP` is
essentially to implement the original `GROUP ... INTO` proposal
and desugar the real `GROUP` into it.

For development and testing purposes, I want it to be possible to
write tests and dev instance queries using the internal `GROUP`,
so this adds syntax for it under the name `FOR GROUP`.

This allows us to reuse with slight modification most of the tests
that are lying around for the old GROUP, so this includes major
updates to that test suite.

This also implement the desugaring (including some optimizations) from
real GROUP into `FOR GROUP`, and updates the toy model to use the
desugaring.