Don't emit the contents of array_agg and the like *twice* (#3665)
Currently we will duplicate an array_agg/array literal in most cases
other than appearing at the very toplevel, once for value and once for
serialized. Avoid this by pretending the serialized output is the value
if we know it is only being used for "not NULL" style things; use eta
expansion to ensure this.