Avoid 'Text file busy' errors during build when CLI is running
You aren't allowed to *modify* a running executable (demand paging
means that .text might not be fully resident in memory, so the kernel
needs to be able to load the correct contents of the executable from
disk), but you can delete it (it will get unlinked from the directory
structure, but a live reference to the file will keep it alive until
the program exits).

So unlink the `edgedb` executable file before doing the copy.