Make CONFIGURE SESSION work in multi-statement execute blocks
We have machinery for splitting out queries that need to be executed
on their own instead of grouped together (`single_unit`), so extend it
to work for CONFIGURE commands and mark CONFIGURE SESSION as
single_unit.

This isn't particularly important, but it is the same infrastructure
that will make SET GLOBAL work in multi-statement execute blocks,
and the error message was wrong in any case.