Fix some issues with defaults
My initial goal here was to delete `ptr_default_to_col_default` and
not set column defaults at all, since we inject the defaults ourselves
anyway. I then realized that the default does serve the actual purpose
of populating data when a new column is created--and that there was a
bug, since multi pointers and things with complex defaults would not
get this treatment, so I used the _alter_pointer_optionality path
to populate that data instead of defaults.

But it also turns out that link properties do actually require having
the defualt column set, since we don't inject the values there. I
didn't want to try to make that change now, so instead I just added an
error for the cases where we can't set a column default on a link
property.

This means I didn't actually accomplish the initial goal, but did fix
a bunch of bugs, so still a win.