update/expand on update example (#3731)
The query given as example on how to update the property of a link needed more clarification to be properly understood. I've also updated the example query to select from the current linked objects, instead of using a detached selection. Both options are described in the comments below the example.