Avoid unnecessary updates to parent views and triggers (#3771)
For parent views, when scheduling inhview updates, instead of finding
all ancestors of the source, find all sources of the ancestors. This
avoids doing silly things like updating BaseObject's view when a
property is added to any type.

For triggers, fix an optimization that avoids updating everything on
changes to inherited links.