Expose Function.backend_name to introspection
This is pretty marginal, but I want it to support some bad scripts I
use to rewrite queries between different instances of a database.

In general, while we don't *support* directly writing SQL queries
against our database, I think that the introspection schema ought
to still contain the information necessary for it to be possible.