Drop broken special case for nested insert FOR (#3797)
There is some code that detects when a nested insert contains a DML
inside an iterator, and explicitly reincludes it. I think that this
was a workaround for when the general DML/iterator mechanisms were
much more broken than they are now.

It causes an extra join, and in the case of multi required things, the
join is broken. Just drop that code.

Fixes #3789.