Drop the never-read-from optional field of qlast.Parameter (#3831)
Properly set cardinality_mod in the enclosing cast, in the locations
it is set. The fixes a bug with OPTIONAL function arguments not having
the correct cardinality. (This was an accident. I was just trying to
delete dead code.)

Also, skip recompiling functions during delete unless it is *actually*
overloaded. This is a minor performance win and also mostly fixes an
issue where it is not possible to delete an existing function that
now fails to compile due to a bug fix.