Fix wrong type name in str->bigint cast error message (#3856)
This turned out to be much more interesting than a simple typo: since
our bigint type is really just a numeric (big *decimal*) with a
constraint on it, if a cast to bigint failed, the postgres error
message would mention numeric, which we would map to std::decimal.

Fix this by catching the error on the cast to numeric and reraising
our own error in that case. Some sources on the internet suggested
this might be slow, so I measured the different versions I tried:
having a separate cast_to_decimal_or_null function that did the error
handling was a slowdown, but inlining that and making the whole thing
plpgsql was actually a 2x speedup.

Fixes #3737.

Also makes a number of the messages more consistent with each other.