Protocol v1.0 - part 1 (#3814)
* Drop support for protocol < 0.13
* Add v0-compatible protocol
* Bump current protocol to 1.0 and preserve compatible support
* Drop statement name of Parse/Prepare message
* Drop statement name of Execute
* Drop the DescribeStatement message
* Rename make_describe_msg to make_command_data_description_msg
* Drop the Execute message

The OptimisticExecute was not ready for the Parse/Execute convention
because it was always looking up in the cache and ignoring the Parse
result. This was an issue because some statement is not cacheable and
compiling again confused the state machine (#3120). Fixed in a way that
the previously parsed result is used if the query hash matches, and
dropped immediately on _execute() to mimic the legacy OptimisticExecute
behavior. test_server_proto_tx_03() and _04() covers both cases.

* Docs: rename Prepare -> Parse, OptimisticExecute -> Execute
* Drop the "optimistic" wording in the code
* Use edgedb 0.24.0a1