Fix a cross-type constraint violation bug
An incorrect optimization would incorrectly decide that an inherited
constraint didn't matter. So in a hierarchy {X, Y} < B < A,
with a constraint on a field in A,
if we try to insert to both X and Y, we will find a constraint
on the nearest common ancestor B and would then decide that the
constraint didn't matter because it was declared not-delegated on
an ancestor, A.

Honestly I can't remember what I was thinking with that logic, which
seems clearly wrong.