Fix a cross-type constraint violation bug (#3887)
An incorrect optimization would incorrectly decide that an inherited
constraint didn't matter. So in a hierarchy {X, Y} < B < A,
with a constraint on a field in A,
if we try to insert to both X and Y, we will find a constraint
on the nearest common ancestor B and would then decide that the
constraint didn't matter because it was declared not-delegated on
an ancestor, A.

It turns out this logic was designed for the case where we also
explicitly considered all ancestors, and so we need to disable it
in the cases where we do not.