Support externally specified object ids (#3895)
Most of the work here is in inserting checks to prevent duplicate
`id`s, since we don't have a trigger on `id` for performance
reasons. We use the conflict mechanism for this. The dodgy bit is that
we have to do it with access policies disabled, to make sure they don't
hide something.

Fixes #3510.