Add support for socket activation (#3899)
Since the idiomatic approach is to have an EdgeDB instance per project
it is easy to end up with dozens of EdgeDB instances running on the user
machine even when idle.  This is problematic from memory consumption
perspective (though a lot less with `on_demand` compiler pool mode).
Thankfully, both systemd and launchd have straightforward support for
socket-activated services.  Combined with auto-shutdown on idle, the
problem of many instances is solved nicely.  This also largely obviates
the need for `--start-conf` on those systems.

Fixes: #3896