Fix a scope leak that caused miscompiles
_try_namespace_fix is too aggressive in obliterating namespaces, and
will do it in cases that causes a reference to a pointer in a
computable to lose its namespacing and improperly correlate with
something outside.

Switch it to using replace_pathid_prefix; this requires fixing
replace_pathid_prefix to not also obliterate namespaces in dodgy ways.

Unfortunately this also breaks a fairly subtle materialization test;
I'm pretty sure that was working by accident, though, and the failure
mode is an ISE, not a miscompilation.

Fixes #3898.