Make Annotation a valid annotation subject
Currently it is accepted in the syntax but then rejected with an ISE.
Fixing this requires a little hackery on the schema side, since it
doesn't support forward references.

Also requires fixing a bug in the schema writer code for AS_LINK,
where the Annotation being updated would be improperly correlated
with the Annotation we are creating a link to.

Also, put std::description annotations on some of the standard
annotations, because now we can.