Fix defaults when doing SET TYPE
If a column has a default, we need to drop the default before doing a
SET TYPE.

Additionally, fix default changing on multi columns.

And while I'm at it, I made it so that we only put defaults on columns
when actually needed: if they are a link property or a sequence.

Fixes #3947.