Fix a collection of group/materialization issues
The core change here is to make setgen.extend_path use a parent ptrcls
when the passed ptrcls is just a trivial derivation.

This is essentially a generalization of the existing 'path_id_name'
mechanism that covers more cases and behaves in a more principled way.
Because we do a better job now of using the right pointer, this allows
eliminating the mechanism where replace_prefix would search for new
pointer refs that directly match the type.

(Eliminating that mechanism was one of the drivers of how this was
done. Despite that all being my code and having *just* gone and
written tests for that behavior, I realized that the whole scheme was
pretty half-baked: calling replace_prefix with prefix==replacement
could cause the path to change!)

Unfortunately the path_id changes regressed
test_edgeql_group_by_group_by_03a, a fairly hairy test that we now
fail to compile.

Closes #3967.
Closes #3955.