Allow multi-statement requests to return data
Currently, if a request contains multiple statements it is considered to
be a "script", i.e. it will not return any data messages back to the
caller.  This is a rather arbitrary limitation at this point, and lifting
it not only simplifies the internal logic, but also enables useful
client patterns, such as DML followed by a `select` without the need to
resort to complex `with` ladders.  Thus, make it so that `output_format`
always controls the output of the trailing statement regardless of how
many preceding statements were in the request (preceding statements
still have their output suppressed).

EdgeQL+HTTP is not supported yet, will be tackled in a followup PR.