Perform HMAC on incoming requests to remote compiler pool server
The remote compiler pool server listens on a TCP port, and the RPC
protocol is currently based on pickle.  These two factors create a
potential for remote code execution if one doesn't run the pool server
in a sandbox.  To be clear, the remote compiler server is an experimental,
untested, and undocumented (aside from appearing in `edgedb-server --help`
currently) feature that isn't part of any current stable release, so the
real-world impact of this code being present in the repository is likely
zero.  Still, it'd be prudent to do the right thing and protect the RPC
with a SHA-256 HMAC using a shared secret.  Non-remote compiler pool uses
a protected Unix socket that is created in a `mkdtemp`-created directory,
so there is no similar concern there.

While at it, hide the remote compiler mode from `--help` to further reduce
the likelihood of somebody actually relying on this mode.

Per report from Elias Hohl.