Implement support for empty ranges
The `std::range()` constructor now has a named only `empty` paramter
that makes for an empty range for a given type.  While its possible to
generate an empty range with equal bounds, an explicit constructor is
a consistent way of doing so in a type-agnostic way, and can also
serve as a canonical output representation of empty ranges.