Fix globals with empty str/bytes values
Currently we implement SET GLOBAL by calling record_send and slicing
off the header and field length, and only storing the encoding
bytes. An empty encoding bytestring is treated as {}. This is
incorrect, because that is the valid encoding of the empty string and
the empty bytes, which are distinguished from NULL by the *length*
field, which is -1 for NULL.

Handle all of this properly by extracting and decoding the length
field and by storing {} values as `None` in the global state, not as
`b''`.