Treat `ABORT MIGRATION` as `ROLLBACK` when in error state
If the corresponding `START MIGRATION TO` started a transaction (as
opposed to a savepoint), then `ABORT MIGRATION` should be allowed to
recover the session like `ROLLBACK`, because that's exactly what it
does.

Also, clean up the relevant code so that it tests for rollback
attributes on the returned query unit instead of going through a
dedicated utility compilation step.  Avoid masking syntax errors and
ISEs with a `TransactionError`.

Fixes: #4056