Make sure that INSERT path ids in ELSE clauses differ from enlcosing
That is, if we have
```
  INSERT Person {name := "Madz"}
  UNLESS CONFLICT ON (.name)
  ELSE (INSERT Person {name := "Maddy"})
```

make sure that the inner `Person` has a distinct pathid (by
namespacing it) from the outer `Person`. They don't correlate despite
the outer one being visible to the inner.

This doesn't fix any current bugs, but that's mostly a matter of
luck. It fixes an issue in an experimental PR I'm working on.